solve crash related to displaying group affiliation was NPE in some cases also ensure chatId is int not long

+++ app/src/main/java/org/awesomeapp/messenger/ui/ConversationListFragment.java
@@ -414,7 +414,7 @@ public class ConversationListFragment extends Fragment {
                     public void onClick(View v) {
                         Context context = v.getContext();
                         Intent intent = new Intent(context, ConversationDetailActivity.class);
-                        intent.putExtra("id", chatId);
+                        intent.putExtra("id", (int)chatId);
                         intent.putExtra("address", address);
                         intent.putExtra("nickname", nickname);
                         context.startActivity(intent);
